refactor: move batch operation related stuff to own package in rdbms
